# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Global Diplomatic Relations platform API
servers:
- url: /api
  description: Base API path
...
paths:
...
  /meetings/{id}/participants:
    get:
      operationId: listMeetingParticipants
      tags:
      - meetings
      summary: List participants for a meeting
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      responses:
        '200':
          description: Participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeetingParticipant'
    post:
      operationId: createMeetingParticipant
      tags:
      - meetings
      summary: Add a participant to a meeting
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingParticipantInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingParticipant'
  /meetings/{id}/participants/{pid}:
    patch:
      operationId: updateMeetingParticipant
      tags:
      - meetings
      summary: Update a meeting participant
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      - $ref: '#/components/parameters/MeetingParticipantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingParticipantUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingParticipant'
        '404':
          description: Participant not found
    delete:
      operationId: deleteMeetingParticipant
      tags:
      - meetings
      summary: Remove a meeting participant
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      - $ref: '#/components/parameters/MeetingParticipantId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Participant not found
  /meetings/{id}/transcripts:
    get:
      operationId: listMeetingTranscripts
      tags:
      - meetings
      summary: List transcripts for a meeting
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      responses:
        '200':
          description: Transcripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeetingTranscript'
    post:
      operationId: createMeetingTranscript
      tags:
      - meetings
      summary: Add a transcript to a meeting
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingTranscriptInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingTranscript'
  /meetings/{id}/transcripts/{tid}:
    patch:
      operationId: updateMeetingTranscript
      tags:
      - meetings
      summary: Update a meeting transcript
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      - $ref: '#/components/parameters/MeetingTranscriptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingTranscriptUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingTranscript'
        '404':
          description: Transcript not found
    delete:
      operationId: deleteMeetingTranscript
      tags:
      - meetings
      summary: Delete a meeting transcript
      parameters:
      - $ref: '#/components/parameters/MeetingId'
      - $ref: '#/components/parameters/MeetingTranscriptId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Transcript not found
...
components:
  parameters:
...
    MeetingId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    MeetingParticipantId:
      name: pid
      in: path
      required: true
      schema:
        type: integer
    MeetingTranscriptId:
      name: tid
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    DeleteResponse:
      type: object
      required:
      - id
      properties:
        id:
          type: integer
...
    MeetingParticipant:
      type: object
      required:
      - id
      - meetingId
      - name
      properties:
        id:
          type: integer
        meetingId:
          type: integer
        contactId:
          type:
          - integer
          - 'null'
        name:
          type: string
        role:
          type:
          - string
          - 'null'
        organization:
          type:
          - string
          - 'null'
        attended:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MeetingParticipantInput:
      type: object
      required:
      - meetingId
      - name
      properties:
        meetingId:
          type: integer
        contactId:
          type: integer
        name:
          type: string
          minLength: 1
        role:
          type: string
        organization:
          type: string
    MeetingParticipantUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
        organization:
          type: string
        attended:
          type: boolean
        contactId:
          type:
          - integer
          - 'null'
    MeetingTranscript:
      type: object
      required:
      - id
      - meetingId
      - authorId
      - authorName
      - content
      - type
      - createdAt
      properties:
        id:
          type: integer
        meetingId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
        type:
          type: string
          enum:
          - transcript
          - notes
          - summary
        createdAt:
          type: string
          format: date-time
    MeetingTranscriptInput:
      type: object
      required:
      - meetingId
      - authorId
      - authorName
      - content
      - type
      properties:
        meetingId:
          type: integer
        authorId:
          type: string
          minLength: 1
        authorName:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        type:
          type: string
          enum:
          - transcript
          - notes
          - summary
    MeetingTranscriptUpdate:
      type: object
      properties:
        content:
          type: string
        type:
          type: string
          enum:
          - transcript
          - notes
          - summary
...